feat(ibl): add indent character
I tried to use the scope only for indenting but with time I discovered
that the scope is not always on. Sometimes it showed the right scope and
sometimes showed the previous scope instead of the new one. This
behiavour made me uncomfortable so I dedicided to enable the indent.